Package Internals

DomainColoring.checkerplotshader — Method

DomainColoring.checkerplotshader(
    w :: Complex;
    real = false,
    imag = false,
    angle = false,
    abs = false,
    polar = false,
)

Takes a complex value w and shades it as in a checker plot.

For documentation of the remaining arguments see checkerplot.

DomainColoring.domaincolorshader — Method

DomainColoring.domaincolorshader(
    w :: Complex;
    abs = false,
    logabs = false,
    grid = false,
    all = false,
)

Takes a complex value w and shades it as in a domain coloring.

For documentation of the remaining arguments see domaincolor.

DomainColoring.labsweep — Method

DomainColoring.labsweep(θ)

Maps a phase angle θ to a color in CIE L*a*b* space by taking

\[\begin{aligned} L^* &= 12 \cos(3\theta - \pi) + 67, \\ a^* &= 46 \cos(\theta + .4) - 3, \quad\text{and} \\ b^* &= 46 \sin(\theta + .4) - 16. \end{aligned}\]

See The Phase Wheel for more information.

DomainColoring.shadedplot — Function

DomainColoring.shadedplot(
    f :: Complex -> Complex,
    shader :: (Complex; kwargs...) -> Color,
    axes = (-1, 1, -1, 1);
    pixels = (720, 720),
    kwargs...
)

Takes a complex function and a shader and produces a Makie image plot.

Arguments

f is the complex function to plot.

shader is the shader function to compute pixel colors.

axes are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the number of pixels to compute in, respectively, the real and imaginary axis, taking the same for both if only one number is provided.

The remaining keyword arguments are passed to the shader function.