Public Interface

DomainColoring.checkerplot — Function

checkerplot(
    f :: "Complex -> Complex",
    axes = (-1, 1, -1, 1);
    pixels = (720, 720),
    real = false,
    imag = false,
    rect = false,
    angle = false,
    abs = false,
    polar = false,
)

Takes a complex function and produces a checker plot as a Makie image.

Arguments

f is the complex function to plot.

axes are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the number of pixels to compute in, respectively, the real and imaginary axis, taking the same for both if only one number is provided.

If none of the below options are set, the plot defaults to rect = true.

real plots black and white stripes orthogonal to the real axis at a rate of 5 stripes per unit.

imag plots black and white stripes orthogonal to the imaginary axis at a rate of 5 stripes per unit.

rect is a shortcut for real = true and imag = true.

phase is a shortcut for angle = true and abs = true.

angle plots black and white stripes orthogonal to the phase angle at a rate of 32 stripes per full rotation.

abs plots black and white stripes at a rate of 5 stripes per unit increase of the natural logarithm of the magnitude.

phase is a shortcut for angle = true and abs = true.

DomainColoring.domaincolor — Function

domaincolor(
    f :: "Complex -> Complex",
    axes = (-1, 1, -1, 1);
    pixels = (720, 720),
    abs = false,
    logabs = false,
    grid = false,
    all = false,
)

Takes a complex function and produces it's domain coloring as a Makie image plot.

Red corresponds to phase $0$, yellow to $\frac{\pi}{3}$, green to $\frac{2\pi}{3}$, cyan to $\pi$, blue to $\frac{4\pi}{3}$, and magenta to $\frac{5\pi}{3}$.

Arguments

f is the complex function to plot.

axes are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the number of pixels to compute in, respectively, the real and imaginary axis, taking the same for both if only one number is provided.

abs toggles the plotting of the magnitude as lightness ramps between level curves.

logabs is similar to abs but shows the natural logarithm of the magnitude instead. This option takes precedence over abs.

grid plots points with integer real or imaginary part as black dots.

all is a shortcut for abs = true and grid = true.

DomainColoring.pdphaseplot — Function

pdphaseplot(
    f :: "Complex -> Complex",
    axes = (-1, 1, -1, 1);
    pixels = (720, 720),
)

Takes a complex valued function and produces a phase plot as a Makie image plot using ColorCET's CBC1 cyclic color map for protanopic and deuteranopic viewers.

Yellow corresponds to phase $0$, white to $\frac{\pi}{2}$, blue to $\pi$, and black to $\frac{3\pi}{2}$.

Arguments

f is the complex function to plot.

axes are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the number of pixels to compute in, respectively, the real and imaginary axis, taking the same for both if only one number is provided.

DomainColoring.tphaseplot — Function

tphaseplot(
    f :: "Complex -> Complex",
    axes = (-1, 1, -1, 1);
    pixels = (720, 720),
)

Takes a complex valued function and produces a phase plot as a Makie image plot using ColorCET's CBTC1 cyclic color map for titranopic viewers.

Red corresponds to phase $0$, white to $\frac{\pi}{2}$, cyan to $\pi$, and black to $\frac{3\pi}{2}$.

Arguments

f is the complex function to plot.

axes are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the number of pixels to compute in, respectively, the real and imaginary axis, taking the same for both if only one number is provided.