DomainColoringToy

DomainColoringToy is an auxiliary package building on DomainColoring.jl which preloads GLMakie and rerenders the plot when the user zooms in or pans around.

The exported functions and arguments are identical to DomainColoring.jl with the addition of the acceptance of :auto in place of an integer in the pixels keyword argument. A direction which is set to :auto will use the viewport resolution to determine the number of samples. Note that this can make plotting very slow.

Installation

DomainColoringToy is a different package and hence has to be installed separately. Installation is as usual:

]add DomainColoringToy

Library

Public Interface

DomainColoringToy.checkerplot — Function

checkerplot(
    f :: "Complex -> Complex",
    limits = (-1, 1, -1, 1);
    pixels = (480, 480),
    real = false,
    imag = false,
    rect = false,
    angle = false,
    abs = false,
    polar = false,
)

Takes a complex function and produces a checker plot as an interactive GLMakie plot.

Arguments

f is the complex function to plot.

limits are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the size of the output in pixels, respectively, the number of pixels along the real and imaginary axis, taking the same for both if only one number is provided. If either is :auto, the viewport resolution is used.

If none of the below options are set, the plot defaults to rect = true.

real plots black and white stripes orthogonal to the real axis at a rate of 5 stripes per unit.

imag plots black and white stripes orthogonal to the imaginary axis at a rate of 5 stripes per unit.

rect is a shortcut for real = true and imag = true.

phase is a shortcut for angle = true and abs = true.

angle plots black and white stripes orthogonal to the phase angle at a rate of 32 stripes per full rotation.

abs plots black and white stripes at a rate of 5 stripes per unit increase of the natural logarithm of the magnitude.

phase is a shortcut for angle = true and abs = true.

DomainColoringToy.domaincolor — Function

domaincolor(
    f :: "Complex -> Complex",
    limits = (-1, 1, -1, 1);
    pixels = (480, 480),
    abs = false,
    grid = false,
    all = false,
)

Takes a complex function and produces it's domain coloring as an interactive GLMakie plot.

Red corresponds to phase $0$, yellow to $\frac{\pi}{3}$, green to $\frac{2\pi}{3}$, cyan to $\pi$, blue to $\frac{4\pi}{3}$, and magenta to $\frac{5\pi}{3}$.

Arguments

f is the complex function to plot.

limits are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the size of the output in pixels, respectively, the number of pixels along the real and imaginary axis, taking the same for both if only one number is provided. If either is :auto, the viewport resolution is used.

abs toggles the plotting of the natural logarithm of the magnitude as lightness ramps between level curves.

grid plots points with integer real or imaginary part as black dots.

all is a shortcut for abs = true and grid = true.

DomainColoringToy.pdphaseplot — Function

pdphaseplot(
    f :: "Complex -> Complex",
    limits = (-1, 1, -1, 1);
    pixels = (480, 480),
)

Takes a complex valued function and produces a phase plot as an interactive GLMakie plot using ColorCET's CBC1 cyclic color map for protanopic and deuteranopic viewers.

Yellow corresponds to phase $0$, white to $\frac{\pi}{2}$, blue to $\pi$, and black to $\frac{3\pi}{2}$.

Arguments

f is the complex function to plot.

limits are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the size of the output in pixels, respectively, the number of pixels along the real and imaginary axis, taking the same for both if only one number is provided. If either is :auto, the viewport resolution is used.

DomainColoringToy.tphaseplot — Function

tphaseplot(
    f :: "Complex -> Complex",
    limits = (-1, 1, -1, 1);
    pixels = (480, 480),
)

Takes a complex valued function and produces a phase plot as an interactive GLMakie plot using ColorCET's CBTC1 cyclic color map for titranopic viewers.

Red corresponds to phase $0$, white to $\frac{\pi}{2}$, cyan to $\pi$, and black to $\frac{3\pi}{2}$.

Arguments

f is the complex function to plot.

limits are the limits of the rectangle to plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

Keyword Arguments

pixels is the size of the output in pixels, respectively, the number of pixels along the real and imaginary axis, taking the same for both if only one number is provided. If either is :auto, the viewport resolution is used.

Package Internals

DomainColoringToy.interactiveshadedplot — Function

DomainColoringToy.interactiveshadedplot(
    f :: "Complex -> Complex",
    shader :: "Complex -> Color",
    limits = (-1, 1, -1, 1),
    pixels = (480, 480),
)

Takes a complex function and a shader and produces a GLMakie plot with auto updating.

Arguments

f is the complex function to plot.

shader is the shader function to compute a pixel.

limits are the initial limits of the plot, in the format (minRe, maxRe, minIm, maxIm), if one or two numbers are provided instead they are take symmetric along the real and imaginary axis.

pixels is the size of the output in pixels, respectively, the number of pixels along the real and imaginary axis, taking the same for both if only one number is provided. If either is :auto, the viewport resolution is used.